requirements.txt
# --- Core Web & Server ---
fastapi==0.135.1
uvicorn==0.41.0
python-multipart
python-dotenv==1.2.2
pydantic==2.12.5
email-validator==2.3.0
pydantic-settings==2.12.0
websockets==15.0.1
httpx==0.28.1

# --- Database & Migrations ---
SQLAlchemy==2.0.48
alembic==1.17.2
psycopg-pool==3.2.7
psycopg-binary==3.2.12
asyncpg==0.30.0

# --- Security & Authentication ---
passlib==1.7.4
bcrypt
PyJWT==2.12.1
argon2-cffi==23.1.0
cryptography==46.0.5
slowapi==0.1.9

# --- AI, LangChain & Embeddings ---
langchain-core==1.2.19
langchain-anthropic==1.3.4
langchain-google-genai==4.2.1
langchain-openai==1.1.11
langgraph==1.1.2
langgraph-checkpoint==3.0.1
langgraph-checkpoint-postgres==3.0.3
langchain-postgres==0.0.16
tiktoken==0.12.0

# --- External Services & Scraping ---
redis==7.3.0
stripe==14.4.1
playwright==1.56.0
playwright_stealth==2.0.2
pdfplumber==0.11.8
google-cloud-storage==3.9.0





